+++ docs/tr2/CHANGELOG.md
@@ -4,6 +4,7 @@
 - added a `/cheats` console command
 - added a `/wireframe` console command (#2500)
 - added a `/fps` console command
+- added `/flood` and `/drain` console commands
 - added support for `-l`/`--level` argument to play a single level
 - added Italian localization to the config tool
 - changed injections to a new file format with a smaller footprint, improved applicability tests and similar feature support as TR1 (#1967)